+++ e2e/esm-config/mjs/__tests__/test.js
@@ -5,10 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {defaults} from '../index';
-
-jest.mock('../importMjs', () => (s: string) => import(s));
-
-test('get configuration defaults', () => {
-  expect(defaults).toBeDefined();
+test('dummy test', () => {
+  expect(1).toBe(1);
 });